requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
werkzeug==2.3.7
apscheduler==3.10.4
gitpython==3.1.40
pathspec==0.12.1
tiktoken>=0.8.0
numpy==1.26.4
rich>=13.7.0
click>=8.1.0
PyYAML>=6.0
google-generativeai>=0.8.0
redis==4.6.0
celery[redis]==5.3.4
minio==7.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
flower==2.0.1
python-json-logger==2.0.7
python-dotenv==1.0.0
psutil==5.9.5